Backend code for Board and Pawn
This file contains BoardC, a representational model of the board. The
board can be called to update which results in row 0 being removed and a
new row 11 being added. There is still no collision detection
implemented. Also during the update, the pieces on the board have a 0.25
chance of moving left or right. Currently there is only one piece being
implemented (Pawn), but I'll implement the others later. This also
contains User which moves left or right.
Excuse all the comments and the console.logs.
This is waiting to be integrated with Ryan's main.