Combined the model and view together
I've combined my backend code with Ryan's visuals to make a working
prototype. The view and model are now in sync allowing the model to
create the visuals for the pieces it creates. Though there is currently
an issue with the piece images overlapping with each other.
I've lowered the game speed for testing purposes.
I've also added fake pause button that stops the animations but doesn't
stop the looping.
Excuse all the comments and console.logs.